requirements.txt
openai
flask
black
python-dotenv